feat: Qual o percentual de obesos entre os homens e entre as mulheres? (É obeso quem tem IMC > 30) check

+++ src/main/java/br/jomoliveira/citel/dtos/PercentualEntreHomensEMulheresDTO.java
@@ -1,4 +1,4 @@
 package br.jomoliveira.citel.dtos;
 
-public record PercentualEntreHomensEMulheresDTO(String sexo, Long populacaoTotal, Long populacaoObesa, Long percentual) {
+public record PercentualEntreHomensEMulheresDTO(String sexo, Long populacaoTotal, Long populacaoObesa, Double percentual) {
 }
